% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_known_peaks.R
\name{detect_peak_centroid}
\alias{detect_peak_centroid}
\title{Detect peak centroid from local spectrum region}
\usage{
detect_peak_centroid(ppm, y)
}
\description{
Detect peak centroid from local spectrum region
}
