% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Perform automated peak detection}
\usage{
detect_peaks(
  spectra.matrix,
  ppm,
  sample.names,
  window.width = "large",
  window.split = 8,
  include.nearbyPeaks = TRUE,
  ...
)
}
\arguments{
\item{spectra.matrix}{A spectral matrix in sample x ppm format}

\item{ppm}{A numeric vector containing the ppm values of spectra.matrix}

\item{sample.names}{A character vector containing the the names each spectrum to perform peak detection within. These should correspond to the rownames of spectra.matrix. By default, peak detection is performed upon all spectra within spectra.matrix.}

\item{window.width}{Specifies the width of the search window during peak detection. See the package speaq for further information.}

\item{window.split}{Specifies the number of units to split the searching window.  during peak detection. See the package speaq for further information.}

\item{include.nearbyPeaks}{Include peaks within the tails of larger neighboring peaks that may be obscured by. If FALSE, these peaks are ignored.}
}
\value{
A dataframe of peaks detected per spectrum and their associated metadata.
}
\description{
This function serves as a wrapper for the speaq::getWavelettPeaks where the default parameters have been optimized for spectra derived from complex media, such as soils. In addition, this function also calculates the width and standard deviation of chemical shift for all peaks to be used in downstream analysis.
}
