% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and filter_spectra.R
\name{extract_spectra_matrix}
\alias{extract_spectra_matrix}
\title{Convert spectra into simplified sample x ppm format}
\usage{
extract_spectra_matrix(NMRData1D_list, samples)
}
\arguments{
\item{NMRData1D_list}{A list of NMRData1D objects}

\item{sample.names}{Character vectors of samples to extract from NMRData1D_list}
}
\value{
A matrix in n x p format, where each row corresponds to a unique spectrum,
columns correpsond to chemical shift and cell values correspond to intesnity
}
\description{
Convert spectra into simplified sample x ppm format
}
