% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess quality of input spectra.R
\name{inspect_raw_spectra}
\alias{inspect_raw_spectra}
\title{Inspect raw and corrected spectra}
\usage{
inspect_raw_spectra(spectra.matrix.raw = NULL, spectra.matrix.corrected = NULL)
}
\arguments{
\item{spectra.matrix.raw}{Raw spectral matrix (samples × ppm).}

\item{spectra.matrix.corrected}{Corrected spectral matrix.}
}
\value{
A Plotly object displaying spectral overlays.
}
\description{
Generates an interactive Plotly overlay of raw and corrected spectra
to visually assess baseline artefacts and correction performance.
}
