% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_known_peaks.R
\name{map_reference_peaks}
\alias{map_reference_peaks}
\title{Map reference peaks to spectra}
\usage{
map_reference_peaks(
  calibrated.output,
  reference.peaks.meta.f,
  spectra.matrix.raw,
  peak.class.map,
  sigma.ppm = 0.01
)
}
\description{
Map reference peaks to spectra
}
