% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peak_assessment}
\alias{peak_assessment}
\title{Assess quality of detected peaks}
\usage{
peak_assessment(
  peak_signals,
  spectra_matrix = NULL,
  ppm = NULL,
  spectra_metadata = NULL,
  treatment = "Landuse",
  treatment_colours = NULL,
  region = NULL,
  outlier_quantile = 0.95,
  min_signal = 3,
  binwidth = NULL,
  legend_position = "right",
  facet_order = NULL
)
}
\arguments{
\item{ppm}{A numeric vector containing the ppm values of spectra.matrix}

\item{treatment}{The name of the corresponding treatment column in spectra.meta}

\item{region}{An optional argument to set the ppm bounds of peaks to inspect}

\item{binwidth}{= An optimal argument used to control the bin width of the diagnostic histogram. By default, this is determined by dividing  the range of SNR values into 50 discrete bins.}

\item{legend_position}{An optional argument used to control the position of the legend in the resulting diagnostic plots.}

\item{peaks.df}{A dataframe of peaks returned by detect_peaks}

\item{spectra.matrix}{The spectral matrix corresponding to peaks.df. If supplied, this matrix used to generate an interactive plot of outlier peaks for each sample.}

\item{spectra.metadata}{A dataframe containing the treatment factors and other metadata associated with each spectrum. If supplied, this is used to determine if there is a significant difference between the composition of peaks among treatment groups.}

\item{treatment.colours}{An optimal list of colours to use in the diagnostic plots}

\item{outlier.quantile}{Numerical threshold used to define the outlier quantile of peaks (default = 0.95)}

\item{min.signal}{The minimum SNR required for a peak to be considered valid (default = 3)}

\item{facet.order}{An optional argument used to specify the order of the resulting diagnostic plots when treatment is supplied.}
}
\value{
A filtered dataframe of valid peaks, alongside several diagnostic plots.
}
\description{
This function assesses the quality of peaks returned by detect_peaks. It returns a filtered dataset of valid peaks, alongside a histogram to help visualize peak quality. This function will also perform a chi-squared test to determine if the distribution of detected peaks across samples (as well as within and between treatment groups if metadata is supplied) deviates from a uniform distribution (i.e. were the detected peaks a result of a true signal, or random noise?)
}
