% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{plot_detected_peaks}
\alias{plot_detected_peaks}
\title{Function to plot detected peaks}
\usage{
plot_detected_peaks(
  df,
  region = NULL,
  samples = NULL,
  reference = NULL,
  spectra_matrix = NULL,
  legend_position = "right",
  metadata = NULL,
  treatment = NULL
)
}
\arguments{
\item{df}{dataframe of grouped peaks to plot}

\item{region}{numeric vector containing range (start,stop) of ppm to plot}

\item{samples}{vector of samples to plot}

\item{reference}{reference spectra used in warping}

\item{spectra_matrix}{spectra matrix used as input for grab_peaks}

\item{legend_position}{control position of plot legend}

\item{metadata}{dataframe contaning metadata of samples}

\item{treatment}{name of treatment column in metadata to group samples by}
}
\value{
Plot of compared spectra
}
\description{
Generate custom plots to vizualise peaks detected by grab_peaks
}
