% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_and_standardisation.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot raw, warped, and standardised spectra}
\usage{
plot_spectra(
  spectra.matrix.raw = NULL,
  spectra.matrix.warped = NULL,
  spectra.matrix.standardised = NULL,
  warping.reference = NULL,
  sample.names = NULL
)
}
\arguments{
\item{spectra.matrix.raw}{Raw spectra matrix.}

\item{spectra.matrix.warped}{Warped spectra matrix.}

\item{spectra.matrix.standardised}{Standardised spectra matrix.}

\item{warping.reference}{Optional reference spectrum.}

\item{sample.names}{Optional subset of samples.}
}
\value{
Named list of Plotly objects.
}
\description{
Generates interactive Plotly spectra comparisons.
}
