% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_known_peaks.R
\name{predict_known_peaks}
\alias{predict_known_peaks}
\title{Predict known peaks using probabilistic assignment}
\usage{
predict_known_peaks(
  total.peaks,
  predicted.known.peaks,
  sigma.ppm = 0.04,
  sigma.auc = NULL,
  use.log.auc = TRUE
)
}
\description{
Predict known peaks using probabilistic assignment
}
