% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and filter_spectra.R
\name{sample_sanity}
\alias{sample_sanity}
\title{Ensure consistency between spectra matrix and metadata}
\usage{
sample_sanity(spectra.matrix, spectra.meta, sample.names = "Spectra.ID")
}
\arguments{
\item{spectra.matrix}{Numeric matrix (samples × ppm).}

\item{spectra.meta}{Data.frame containing metadata.}

\item{sample.names}{Column name in metadata containing sample IDs.}
}
\value{
List containing filtered spectra.matrix and spectra.meta.
}
\description{
Filters both spectral matrix and metadata so that only shared sample IDs remain.
}
