% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_known_peaks.R
\name{summarise_reference_peaks}
\alias{summarise_reference_peaks}
\title{Summarise reference peaks across spectra}
\usage{
summarise_reference_peaks(
  reference.peaks.meta.f,
  spectra.matrix.raw,
  peak.class.map,
  ppm.window.base = 0.04,
  smooth = TRUE,
  k.window = 3
)
}
\description{
Summarise reference peaks across spectra
}
